Make immediate test test interleaving.
Which, as per last commit, I'm not sure is really totally fair? I mean...
I think it might be... there ought be a flush of the task queue that ping
pongs, I think, between the two readIntos, but I'm not fully sure that's
absolutely positively the case, but I do think it might be. Anyhow, like
this test for now. Wfm?